Fix race condition when creating a VM.
This change removes the threading.Lock() and replaces it with an asyncio.Lock(), which
was necessary because the threading.Lock() is ineffective in the context of asynchronous code.

Other parts of the code have been amended to use less synchronous code.  However, the API
of the `Volume` class itself has synchronous public methods, which are a pain to work around
-- as a result, I have had to remove some caching and this will run a bit slower.