Ensure revert works as expected.
An expectation of qvm-volume revert is that, upon rollback, what is recalled is the latest revision.  Now, formerly, in the ZFS driver, we snapshot a revision when a VM stopped, which means the latest revision was always identical to the volumes state after poweroff, which meant reverting was a noop, always.

In this PR, we no longer use revisions as clone sources.  Instead, we have our own clean snapshots, not shown to the revision machinery, to perform clones from.

With this mechanism, we can now finally take a revision *before VM start*, which means that reversion to the latest snapshot actually reverts to a prior state, instead of being a noop.  This is nice.

By default snapshots used for revisions are deleted when the file system is rolled back to them (qvm revert).  There is a boolean that controls that behavior.

Ref: https://github.com/QubesOS/qubes-core-admin-linux/pull/118#issuecomment-1457938151