Address bug introduced with the atomic creation patch.
The atomic creation mechanism is as follows:

* Clone (from ZFS source) or create (for `dd of=`) a temporary volume.
  * This volume is stored in `pool/.tmp/name-of-dataset`
* Ensure clone is finished.
* Swap the dataset into place.
  * Its final place will be `pool/dataset/{private,root,volatile...}`

The problem with this algorithm is that `pool/dataset` must exist
before `pool/.tmp/name-of-dataset` can be atomically renamed to
`pool/dataset/X`.

The solution is to add `-p` flag to the rename call.

Some cache optimizations are also included here.

Finally, a detection of whether the pool of the source and the pool
of the target has been added.  Without this, cross-pool cloning
would just fail because the source pool and the target pool were
different, and ZFS clone cannot clone across pools.